remove the use of datahelper from InstallSchema
Problem:

When we do a clean magento2 install we get an exception caused by this
module.

  [Magento\Framework\Exception\LocalizedException]
  Invalid entity_type specified: customer

The reason for this is that de InstallData is executed after all
installations of schemas and updates of schemas.

Solution:

Remove the use of the helper in the InstallSchema, the helper will cause
the ObjectManager to instantiate several objects and that causes the
exception because the eav entity type for customer is not yet set.

The helper got removed from the InstallSchema, and due to that we use BP
now instead of the helpers getBasePath method. Additionally the try
catch to ignore the failure of creating the MailChimp folder in var is
removed. I think its better to fail hard there instead of ignoring the
fact the folder could not be created.

Signed-off-by: BlackEagle <ike.devolder@gmail.com>